Added biased random walks
I modified the Random_Walk class so that it now supports biased random walks (in addition to symmetric ones).

To do this, I added a second argument to the Random_Walk constructor. This argument, called p, is default set to 1/2. The user can therefore optionally supply another value of p to create a biased random walk.

With this, I wrote code for problem 3.6.4. part 3 (which used a biased random walk).

I implemented the biased random walk in such a way that it does not change the performance of the code. To do this, I got rid of the HALF_RAND_MAX variable and replaced it with p_RAND_MAX. This new variable is simply p times RAND_MAX. I then compare x (the random value) to p_RAND_MAX rather than HALF_RAND_MAX (See Random_Walk.cc:37).

This code runs and has been tested.